webpage/requirements.txt
flask==2.0.1
flask-wtf==0.15.1
discord.py==2.0.0
requests==2.26.0
python-dotenv==0.19.0 